Fix typo and readme bug
On line 70 the production server can not find the bundle.js
file because webpack.config.js pushes the bundle.js into the
/public path. The result is a blank page because the
server has not been set up with the proper path of /public.
The Webpack configuration must be applied after the public
directory has been created on line 80 or once the server has
been configured to read from the public directory.